requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
dnspython==1.16.0
idna==2.8
iexfinance==0.4.2
numpy==1.17.0
optparse-pretty==0.1.1
pandas==0.25.0
pymongo==3.9.0
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.3
urllib3==1.25.3
torch==1.3.1
matplotlib==3.1.2